Change ConfirmCompositionText() from uint32_t to size_t.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3783487
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1028922}

+++ components/exo/text_input.cc
@@ -189,14 +189,13 @@ void TextInput::SetCompositionText(const ui::CompositionText& composition) {
   delegate_->SetCompositionText(composition);
 }
 
-uint32_t TextInput::ConfirmCompositionText(bool keep_selection) {
+size_t TextInput::ConfirmCompositionText(bool keep_selection) {
   // TODO(b/134473433) Modify this function so that when keep_selection is
   // true, the selection is not changed when text committed
   if (keep_selection) {
     NOTIMPLEMENTED_LOG_ONCE();
   }
-  const uint32_t composition_text_length =
-      static_cast<uint32_t>(composition_.text.length());
+  const size_t composition_text_length = composition_.text.length();
   delegate_->Commit(composition_.text);
   ResetCompositionTextCache();
   return composition_text_length;
